SPARC:
So I finally worked out the most convoluted bug I ever produced. Turns
out there were too many nops after re-enabling traps during the thread
switch sequence, which caused the store instruction of the old thread's PSR
to be preempted. Since the value of the %psr is stored in a global
register, that register was overwritten with another PSR-value during
the switch sequence. The register windows of the old thread are flushed just
before the switch, thereby (potentially) storing only the %i and %l registers
to the memory stack in case a window overflow trap occurs. However, in
case of a more complex trap type (such as an interrupt), ALL the
registers are stored into a special stack frame. In this frame, the
storage location of the last frame's %g4 and %i6 (i.e. %fp) are at
identical offsets in memory.

Now, in the rare occasion where 1) an old thread returned, thereby forcing
a scheduling event coincided with 2) the scheduling timer triggering an
IRQ and 3) the IRQ happened to exactly preempt the current code path in
the location of the store of %g4 (containing the PSR) to the thread-info
struct and 4) the preempted thread being the same which would have been
scheduled following the IRQ, the value at %g4 in memory of the "simple"
stack-save (i.e. of %i and %l registers) would be read as the value of %i6
following the restore sequence in the trap return path, which would
become %o6 (aka %sp) after finally returning from the trap.

The fix itself is simple: just eliminate enough nops so that toggling
the ET bit in the %psr would not take effect until after the old
thread's %psr is stored.